refactor(upstream_config): split into focused submodules
Break the 1561-line src/aggregator/upstream_config.rs into a directory
module with cohesive submodules:

- mod.rs        — config types, ConfiguredUpstreams/ProviderSessionRegistry,
                  UpstreamSessionSink, UpstreamConfigManager
- validation.rs — config file parse/validate, verification-target derivation
- builders.rs   — backend/verifier construction from config
- dynamic.rs    — Empty/Dynamic backend + verifier wrappers
- tests.rs      — the existing unit tests

Pure relocation: no logic changes. Public paths preserved
(parse_config_text re-exported); cross-module helpers marked pub(super).
cargo test --all-targets: 220 passed, 0 failed (unchanged from baseline).
